Remove watching of execution plan
Unless delegating the computation to some other service, there is
quite often need for running some tasks for some time, if fact, that's
one of the reasons why we use background processing. There should be a
way how to terminate some task if running too long, but 1 second is
not the case anyway.