Extract consistency check outside of executor
The consistency check should not be buried into parallel executor.
Ideally, it should not be dependent on no executor running (should be
solved with register of executors). But for now, I think it's better
to leave the responsibility of running the consistency check the right
moment for the Dynflow user.

With the executors register in place, it should even not matter when
or where the consistency check is run.

This also allows us getting rid of the @world.initielized future.