Fix tests on random unresolved future
There was a case where an event came right at the time of the core
being terminated, and ended up in the unresolved state.

Handling this in actor by dead letter routing is not as easy as it
seems (as making sure to read all the messages after termination
turns out to not be so straight forward). So for now we just make sure
to refuse all the incoming events as soon as possible.