Sub plans support
This patch adds support for triggering execution plans from a run phase
(a.k.a sub-plans) and waiting for the sub-plans to finish.

It's as simple as:

    class ActionWithSubPlan < Dynflow::Action
      include Dynflow::Action::WithSubPlans

      def create_sub_plans
        10.times.map { trigger(SomeAction, input) }
      end
    end

The `Action::WithSubPlans` defines the `run` method and
expects the action to define `create_sub_plans` to return
one or many triggered actions. The `WithSubPlans` module takes
care of waiting for the sub plans to finish, and succeeding
or failing after that.

The resuming behaviour can be also customized, by overriding
the `resume` method. By default, it re-creates the sub plans
in case none of the sub plans was planned successfully. Otherwise
it checks, that all the sub plans have already finished and fails
when they haven't.

To support this, the caller_execution_plan_id and caller_action_id attributes
were added to the action, to track this relation. The relation between actions
in one execution plan is tracked the same way as well, with the prospect of
using this data instead of the data in execution plan later (but we are not
there yet).