Set logger before validating world
Currently when a Rails application initializes a new world, Dynflow may
use the logger from that world on the validation. However, that fails
because the logger hasn't been defined until the config for the wolrd
has been validated.

See how Config#validate calls 'config_for_world.world.logger'. That will
in turn call World#logger, which calls `logger_adapter` and fails
because the logger has not been set yet.

This commit ensure that there is a logger available every time we
initialize a world, therefore being able to validate the config on Rails
apps instead of failing