Refs #17283 - Extract decision making about what to do from the executor
Before patch patch, we had the logics about what to execute hidden
inside the parallel executor. In order to reuse as much as possible in
prepared in-thread executor, as well as making better separation of
responsibilities, the parallel executor is now responsible only for the
execution itself, delegating the selection of the work items on newly
created director.