Fixes #19146 - handle invalid plans when invalidating worlds
In a corner case, the data can be inconsistent in a way, that prevents
the world to be invalidate properly. Although this is usually a result
of unsafe manipulation with data on database level, we need to try
to resolve the situation to at least keep clean the runtime data properly,
as there is not much more we can do about it.