Fix compatibility with algebrick 0.7.4
Algebrick 0.7.4 defines eql? and hash function on the objects. The
problem is that in clock, the receiver is part of the algebrick object,
which leads to the receiver to be included in the hash calculation. What
can happen is the application of the timer actually modifies the
receiver so that it alters the hash value of the object, and it failed
to delete the object from list of timers, leading to endless loops.

The most simple example is this:

   clock.ping([], 0.1, :pong, :<<)
   sleep 0.2

Once the ping realizes, the receiver changes from `[]` to `[:pong]` and
it fails to delete the timer from the list.

Fallbacking to using `object_id` as the hash function should work for us
more safely for this particular use-case.