Fixes #20647 - fix calculation of rescue strategy for complex actions
Since ba6300f, we were considering also pending and successful steps
when calculating the rescue strategy. This made the auto_rescue not
working properly, as any action with `Pause` rescue strategy (which is
considered by default) would lead to pause of the task.

This change makes the rescue strategy working properly again, while
keeping the Fail strategy working as well.