Fixes #20648 - support for multiple auto-rescues in the plan
Currently, when an error occurs and skip rescue strategy applies,
it auto-rescues only at the first failure: the auto-rescue is not
applied for further failures.

After this change, we don't check only if the plan was already rescued
or not, but we check, if there is some failed action that we have not
tried to rescue from yet: this is better check to ensure we don't
run into endless loop, while allowing to auto-rescue several times
in a row.