Introduce Singleton action module
An action into which this module is included (singleton action from
now on) on which there should be only one instance running at any
given time.

This means the action tries to obtain a lock at the beginning of its
plan phase and fails if it doesn't acquire it. At each Action#run it
checks if it is still holding the lock.  If it does, it goes on,
otherwise tries to reacquire it. It behaves the same way during the
lock reacquiring as in plan phase.

Finally, it releases the lock at the end of finalize phase. If
things go wrong, locks should get released when the execution plan
holding them gets invalidated.

Please make sure singleton actions release the lock they're holding
in case they encounter any error from which they can't recover.