Cache Ping messages
Some other projects use Pings quite extensively (run them at the start
of every actions) so it seemed like a good idea to cache them to reduce
the number of sent messages.

This commit introduces such mechanism. Cache records are added either
when a response to a Ping or a Ping is received. This way, both
participants in the communication have the other one in their own
caches.

Furthermore a world tries to announce its availability (pings all other
worlds) at the end of its #initialize.