Fix intermittent failures: abnormal_states_recovery_test
Wait until the step gets suspended: as the executing state of the
execution plan still can be sooner than the executor started the work.

I was running the tests in a loop against postgres database to reproduce
the issue:

    export DB_CONN_STRING='postgres://postgres@/dynflow_example'
    I=0; while bundle exec ruby -I"lib:lib:test" "test/abnormal_states_recovery_test.rb" -n '/plans as paused/'; do I=$((I + 1)); echo "========== $I"; done

I usually got the failure in 20 iterations, after the fix, I have not
seen the issue so far.