First draft of execution plan hooks.
This commit allows actions to hook into the life cycle of an execution plan.
Currently hooks can be triggered when the execution plan goes into:
- stopped state (:stop)
- paused state (:pause)
- stopped state with success result (:success)
- stopped state with error, warning or canceled result (:fail)

A hook can either override default `execute` method and provide the same
behavior for all events or override separate `on_success`, `on_stop`,
`on_pause` and `on_fail` methods to implement the desired behavior.

Hooks have volatile access to the action, any changes done to the action's
output or input will be lost upon exiting the hook.

Hooks are enabled using the `execution_plan_hooks.use` call and are inherited
when actions are sub-classed. Hooks from the parent action can be disabled using
the `execution_plan_hooks.do_not_use` call in the child action. The order of
execution is not guaranteed.

The hooks are executed in a begin-rescue block, so even if a hook raises an
exception during its run, other hooks should still be run.