Implement executor heartbeat
Executors now do a heartbeat every 15 seconds, freshness interval was
increased to 60 seconds. On heartbeat executors update the world's
last_seen, which is then used to determine whether a world is valid or
not.

Because only executors perform the heartbeat, client worlds still need
to have Pings delivered to them.