Reduce status updates on rescue
The previous rescue algorithm put the task into paused state and then we
looked at what are the possible rescue strategies, and if it was to
skip, we marked the steps to skip and executed the task again.

Several problems where there with this approach:

- the execution plan needed to get back at the beginning of the workers
  queue, so there was additional overhead with this

- the execution history got confusing, because there was note about
  finishing and starting the execution again

- it's hard to hook events (such as logging or notifications) for paused
  tasks, as when the plan was getting into paused state, we didn't know
  it would be rescued immediately.

This is a prerequisite for allowing notifications on paused tasks.