Prepare the EventWorkItem to be serializable
Before this patch, the event object within `EventWorkItem` included
a `resolvable_future`, that could not be serialized (which we need for
being able to send the work item to sidekiq.

Because of this, we introducing unique `request_id` for the event and
using it to track the result of the execution back to the resolvable
future that we are expected to fulfill.

We also needed to postpone creation of event work items to time of other
work items for the step being processed: the reason is the fact that we
need to update the step based on the latest result from the executor
first, otherwise the step status we would have would not be updated and
we would get error of trying to run event on non-suspended state.