Allow running multiple orchestrators in active-passive mode (#335)
Upon startup, the orchestrator checks if dynflow_orchestrator_uuid key
is unset in Redis. If so, it acquires an orchestrator lock by
associating the uuid of the orchestrator's world with the key. However
this lock has limited time to live and needs to be periodically
refreshed.

If the key is set, the orchestrator goes into passive mode and
periodically polls the lock's status. If it manages to acquire the lock,
it becomes the active orchestrator.

Active orchestrator periodically updates the record in Redis to ensure
its freshness.