Handle cases where worker is shut down
If the worker is shut down, execution of its work item is interrupted, but it
still sends the work done job to the orchestrator and requeues the job to be
executed again. This leads to two issues - step is not saved after the worker is
interrupted and the job is executed twice.

This patch disables the job requeueing and makes orchestrator save the step, if
the step's error is ::Sidekiq::Shutdown. This solves both of the issues.

If the worker is not interrupted, but is SIGKILLed, then the regular worker
crash handling kicks in, reruns the job and the it fails because the step is in
an unexpected state.