Fix how we determine AR connection pool size when using Sidekiq
Before Sidekiq we configured the AR connecton pool size as sum of all configured
workers + 5. When using Sidekiq we should take the number of worker threads
Sidekiq has and add some constant to it, instead of relying on queue definitions
which don't have any impact on number of threads when running under Sidekiq.