Deal with AR connection pool shenanigans
The behavior around cleanup of active record connections is driven by several
entry (before user code gets executed) and exit (after user code gets executed)
conditions.

entry-A: The thread currently does not have any connection from the ActiveRecord
connection pool.

entry-B: The thread has a connection from ActiveRecord connection pool and there
are no opened transactions associated with the connection.

entry-C: The thread has a connection from ActiveRecord connection pool and there
is at least one opened transaction already associated with that connection.

After yielding and executing the client code, there are two options:

exit-A: The user code did not need an ActiveRecord connection.

exit-B: The user code used ActiveRecord and either obtained the connection from
the pool (entry-A) or the thread already had a connection (entry-B or entry-C).
Either way, the thread has a connection after executing the user code.

Now we're finally getting to the cleanup. It makes sense to perform the cleanup
if it is safe to do so (entry-A or entry-B) and if the current thread has a
connection ((entry-A and exit-B) or entry-B).

That leaves us with (entry-A and exit-A) where no cleanup is necessary and
entry-C where cleanup is not safe to do. We used to perform the cleanup every
time (including entry-C) and some connection errors came from that because we
returned the connection in the middle of a transaction.