Do not instantiate an exception to check its class
When the worker is stopped, ::Sidekiq::Shutdown exception is raised
inside the worker. In case that happens, we need to save the step on
from the orchestrator's side to not lose data and progress. That means
we have to check every single step if it has an error and if so, what
kind of error it was. Originally we checked whether the exception is
an instance of ::Sidekiq::Shutdown. This commit changes this behavior
to not check the instance, but the class itself.

This may not work if the class is a subclass of the class we're
checking for, but given we expect one specific exception class, I'd
say this is safe to do.