Implement MQTT based connector
Historically Dynflow used database for communication across processes. The
sending side would write a record into dynflow_envelopes table, receiving side
would poll this table and if it found a record addressed to it, it would receive
it and remove it from the db. Both polling and insert-read-delete cycles were
putting unnecessary load on the db. Using a message bus for this should be the
way to go.

Here I chose MQTT because it is lightweight and I worked with it recently and
still have it in fresh memory. However, this could be adjusted to almost
anything with minimal changes.