Add support for chunked outputs
If an action suspends and resumes a lot and produces a lot of "log style"
output, it makes execution slow. Appending a single line to an array in output
means Dynflow has to load the entire output, deserialize it, append the line,
serialize it and then write to a database. For small outputs this is fine, but
it doesn't scale well.

This commit introduces support for append-only outputs. Such outputs can be only
appended or retrieved as a whole, but cannot be updated in place.