Add execution plan chaining
This commit enables execution plans to be chained. Assuming there is an
execution plan EP1, another execution plan EP2 can be chained onto EP1.  When
chained, EP2 will stay in scheduled state until EP1 goes to stopped state. An
execution plan can be chained onto multiple prerequisite execution plans, in
which case it will be run once all the prerequisite execution plans are stopped
and have not failed. If the prerequisite execution plan ends with
stopped-error, the chained execution plan(s) will fail. If the prerequisite
execution plan is halted, the chained execution plan(s) will be run.

It builds on mechanisms which were already present. When an execution
plan is chained, it behaves in the same way as if it was scheduled for
future execution. A record is created in dynflow_delayed_table and once
the conditions for it to execute are right, it is dispatched by the
delayed executor. Because of this, there might be small delay between
when the prerequisites finishs and the chained plan is started.